FCOS/requirements.txt
Keras==2.2.5
keras-resnet==0.2.0
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==6.2.0
tensorflow-cpu==1.15
progressbar2
